Determine source and reflection regions, max distances, etc
Reflection matching is only useful in a certain a range of distances. We
calculate the maximum distance as the point where the slope of the disparity
function goes to -1. This is somewhat arbitrary, but ok for now. The min
distance is provided by the user.

Using these values, we can find a region that is mostly source, and a region
that is mostly reflection. There is also a third region where it is not certain.
These are now bounded by red lines in the image.